Renamed module from lws to websocket
