+++ modules/websocket/register_types.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  register_types.cpp                                                   */
+/*  register_types.h                                                     */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -27,53 +27,5 @@
 /* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE     */
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
-#include "register_types.h"
-#include "error_macros.h"
-#ifdef JAVASCRIPT_ENABLED
-#include "emscripten.h"
-#include "emws_client.h"
-#include "emws_peer.h"
-#include "emws_server.h"
-#else
-#include "lws_client.h"
-#include "lws_peer.h"
-#include "lws_server.h"
-#endif
-
-void register_lws_types() {
-#ifdef JAVASCRIPT_ENABLED
-	EM_ASM({
-		var IDHandler = {};
-		IDHandler["ids"] = [];
-		// get an integer ID for a JS object. this keeps a reference to it, preventing GC'ing
-		IDHandler["add"] = function(obj) {
-			var id = IDHandler.ids.length;
-			IDHandler.ids[id] = obj;
-			return id;
-		};
-		// get a JS object from an integer ID
-		IDHandler["get"] = function(id) {
-			return IDHandler.ids[id];
-		};
-		// releases an object that has an ID. this allows it to be GD'd
-		IDHandler["remove"] = function(id) {
-			IDHandler.ids[id] = null;
-		};
-		Module["IDHandler"] = IDHandler;
-	});
-	EMWSPeer::make_default();
-	EMWSClient::make_default();
-	EMWSServer::make_default();
-#else
-	LWSPeer::make_default();
-	LWSClient::make_default();
-	LWSServer::make_default();
-#endif
-
-	ClassDB::register_virtual_class<WebSocketMultiplayerPeer>();
-	ClassDB::register_custom_instance_class<WebSocketServer>();
-	ClassDB::register_custom_instance_class<WebSocketClient>();
-	ClassDB::register_custom_instance_class<WebSocketPeer>();
-}
-
-void unregister_lws_types() {}
+void register_websocket_types();
+void unregister_websocket_types();
